requirements.txt
# AtlasOps Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# HTTP Client
httpx>=0.26.0

# Task Queue
celery[redis]>=5.3.6
redis>=5.0.1

# AI/LLM
openai>=1.10.0

# Web Scraping
beautifulsoup4>=4.12.3
playwright>=1.41.0

# PDF Generation
weasyprint>=60.2

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0

# Development
ruff>=0.1.14
black>=24.1.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
